Add rejoin logic
De-activate reports when a user is no longer in the guild. Re-activate if the user rejoins the guild.

On rejoin, send an embed into the report channel stating that the user has returned. Re-apply a mute if the user is detained.

Add a reaction to messages to denote failed actions.