Fix migration bugs
Fixed a bug where migration embeds were sent in a report when a user with a report migrated in/out of a different server than the one the report was opened in.
Fixed an exception that was thrown on a user leaving the guild and having messages sent in their report. This now gives the proper fail reaction response.
Fixed a logic error where channels deleted manually would not close properly. This error was added upon removal of the AuditLogPollingService in a previous commit, not in a released version.